+++ demos/browser/app/meetingV2/meetingV2.ts
@@ -237,12 +237,6 @@ export class DemoMeetingApp
   static readonly DATA_MESSAGE_TOPIC: string = 'chat';
   static readonly DATA_MESSAGE_LIFETIME_MS: number = 300_000;
 
-  // Currently this is the same as the maximum number of clients that can enable video (25)
-  // so we id the check box 'enable-pagination' rather then 'reduce-pagionation', but technically pagination is always enabled.
-  static readonly REMOTE_VIDEO_PAGE_SIZE: number = 25;
-  // Enabled on authentication screen by 'enable-pagination' checkbox.
-  static readonly REDUCED_REMOTE_VIDEO_PAGE_SIZE: number = 2;
-
   // Ideally we don't need to change this. Keep this configurable in case users have a super slow network.
   loadingBodyPixDependencyTimeoutMs: number = 10_000;
   loadingBodyPixDependencyPromise: undefined | Promise<void>;
@@ -334,6 +328,7 @@ export class DemoMeetingApp
   voiceFocusDevice: VoiceFocusTransformDevice | undefined;
   joinInfo: any | undefined;
   deleteOwnAttendeeToLeave = false;
+  disablePeriodicKeyframeRequestOnContentSender = false;
 
   blurProcessor: BackgroundBlurProcessor | undefined;
   replacementProcessor: BackgroundReplacementProcessor | undefined;
@@ -570,12 +565,15 @@ export class DemoMeetingApp
       const priorityBasedDownlinkPolicyConfigTitle = document.getElementById(
           'priority-downlink-policy-preset-title'
       ) as HTMLElement;
-      const enablePaginationCheckbox = document.getElementById(
-          'enable-pagination-checkbox'
-      ) as HTMLSelectElement;
       const serverSideNetworkAdaption = document.getElementById(
           'server-side-network-adaption'
       ) as HTMLSelectElement;
+      const paginationPageSize = document.getElementById(
+        'pagination-page-size'
+      ) as HTMLElement;
+      const paginationTitle = document.getElementById(
+        'pagination-title'
+      ) as HTMLElement;      
       const serverSideNetworkAdaptionTitle = document.getElementById(
           'server-side-network-adaption-title'
       ) as HTMLElement;
@@ -583,14 +581,16 @@ export class DemoMeetingApp
       if (this.usePriorityBasedDownlinkPolicy) {
         priorityBasedDownlinkPolicyConfigTitle.style.display = 'block';
         priorityBasedDownlinkPolicyConfig.style.display = 'block';
-        enablePaginationCheckbox.style.display = 'block';
         serverSideNetworkAdaption.style.display = 'block';
+        paginationPageSize.style.display = 'block';
+        paginationTitle.style.display = 'block';
         serverSideNetworkAdaptionTitle.style.display = 'block';
       } else {
         priorityBasedDownlinkPolicyConfigTitle.style.display = 'none';
         priorityBasedDownlinkPolicyConfig.style.display = 'none';
-        enablePaginationCheckbox.style.display = 'none';
         serverSideNetworkAdaption.style.display = 'none';
+        paginationTitle.style.display = 'none';
+        paginationPageSize.style.display = 'none';
         serverSideNetworkAdaptionTitle.style.display = 'none';
       }
     });
@@ -1717,6 +1717,7 @@ export class DemoMeetingApp
       configuration.videoDownlinkBandwidthPolicy = this.priorityBasedDownlinkPolicy;
       this.priorityBasedDownlinkPolicy.addObserver(this);
     }
+    configuration.disablePeriodicKeyframeRequestOnContentSender = this.disablePeriodicKeyframeRequestOnContentSender;
 
     configuration.applicationMetadata = ApplicationMetadata.create('amazon-chime-sdk-js-demo', '2.0.0');
 
@@ -1763,10 +1764,13 @@ export class DemoMeetingApp
       this.audioVideo.setVideoCodecSendPreferences(this.videoCodecPreferences);
       this.audioVideo.setContentShareVideoCodecPreferences(this.videoCodecPreferences);
     }
+
+    // The default pagination size is 25.
+    let paginationPageSize = parseInt((document.getElementById('pagination-page-size') as HTMLSelectElement).value)
     this.videoTileCollection = new VideoTileCollection(this.audioVideo,
         this.meetingLogger,
         this.usePriorityBasedDownlinkPolicy ? new VideoPreferenceManager(this.meetingLogger, this.priorityBasedDownlinkPolicy) : undefined,
-        (document.getElementById('enable-pagination') as HTMLInputElement).checked ? DemoMeetingApp.REDUCED_REMOTE_VIDEO_PAGE_SIZE : DemoMeetingApp.REMOTE_VIDEO_PAGE_SIZE)
+        paginationPageSize)
     this.audioVideo.addObserver(this.videoTileCollection);
 
     this.contentShare = new ContentShareManager(this.meetingLogger, this.audioVideo, this.usingStereoMusicAudioProfile);
@@ -3356,6 +3360,7 @@ export class DemoMeetingApp
     this.enableSimulcast = (document.getElementById('simulcast') as HTMLInputElement).checked;
     this.enableEventReporting = (document.getElementById('event-reporting') as HTMLInputElement).checked;
     this.deleteOwnAttendeeToLeave = (document.getElementById('delete-attendee') as HTMLInputElement).checked;
+    this.disablePeriodicKeyframeRequestOnContentSender = (document.getElementById('disable-content-keyframe') as HTMLInputElement).checked;
     this.enableWebAudio = (document.getElementById('webaudio') as HTMLInputElement).checked;
     this.usePriorityBasedDownlinkPolicy = (document.getElementById('priority-downlink-policy') as HTMLInputElement).checked;
     this.echoReductionCapability = (document.getElementById('echo-reduction-capability') as HTMLInputElement).checked;