+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.11.0] - 2023-01-03
 
 ### Added
+- Add `meetingComponentLoadTime` to event controller's `EventName` type.
 
 ### Removed